EquipGroups finished!
Any item will holdable (not yet!). You can only access it's
"equipActions" (e.g. shoot gun) if it is equipped in the slots its
EquipGroups define. Example: You can pick up a rifle with your right
hand, but not shoot it. In order to shoot it you must equip it to both
hands (RightHeld and LeftHeld), where you'll gain access to "Fire".